16_51 uodate arreglado, se implemento agenda medic

+++ services/file_service.py
@@ -1,46 +1,34 @@
-from datetime import datetime, timedelta
 from repositories.file_repository import FileRepository
 from models.file import File
-from models.patient import Patient
-from models.medic import Medic
+from datetime import datetime, timedelta
 
 class FileService:
 
     @staticmethod
-    def create_file(patient, medic_id, details, presentation_time):
-        medic = Medic.query.get(medic_id)
-        if not medic:
-            return {'message': 'Medic not found'}, 404
-
+    def create_file(patient_id, medic_id, details, presentation_time_str):
+        presentation_time = datetime.fromisoformat(presentation_time_str)
+        creation_date = datetime.utcnow()
         new_file = File(
-            patient_id=patient.id,
-            medic_id=medic.id,
+            patient_id=patient_id,
+            medic_id=medic_id,
             details=details,
-            presentation_time=datetime.strptime(presentation_time, '%Y-%m-%dT%H:%M:%S')
+            state='pendiente',
+            creation_date=creation_date,
+            presentation_time=presentation_time
         )
-
         FileRepository.save(new_file)
+        return {'message': 'File created successfully'}, 201
 
-        return {
-            'Patient': f"{patient.last_name} {patient.name}",
-            'Medic': f"{medic.last_name} {medic.name}",
-            'Specialty': medic.speciality,
-            'Presentation Time': new_file.presentation_time,
-            'Creation Date': new_file.creation_date,
-            'Details': new_file.details,
-            'State': new_file.state
-        }, 201
-    
     @staticmethod
     def get_all_files():
         files = FileRepository.get_all()
         return [{'id': fl.id, 'patient_name': f"{fl.patient.name} {fl.patient.last_name}", 
-                 'medic_name': f"{fl.medic.name} {fl.medic.last_name}",
-                 'speciality': fl.medic.speciality, 
-                 'presentation_time': fl.presentation_time, 
-                 'creation_date': fl.creation_date, 
-                 'details': fl.details, 
-                 'state': fl.state} for fl in files], 200
+                'medic_name': f"{fl.medic.name} {fl.medic.last_name}",
+                'speciality': fl.medic.speciality, 
+                'presentation_time': fl.presentation_time, 
+                'creation_date': fl.creation_date, 
+                'details': fl.details, 
+                'state': fl.state} for fl in files], 200
 
     @staticmethod
     def get_file_by_id(file_id):
@@ -58,38 +46,53 @@ class FileService:
             }, 200
         return {'message': 'File not found'}, 404
 
-
-
-
-
-
-
-
     @staticmethod
-    def update_file(file_id, data):
+    def update_presentation_time(file_id, new_presentation_time_str, current_user):
         file = FileRepository.get_by_id(file_id)
         if not file:
             return {'message': 'File not found'}, 404
 
-        patient = Patient.query.get(file.patient_id)
-        if datetime.utcnow() > file.creation_date + timedelta(minutes=15):
-            return {'message': 'Cannot update file after 15 minutes of creation'}, 403
+        new_presentation_time = datetime.fromisoformat(new_presentation_time_str)
 
-        if 'details' in data:
-            file.details = data['details']
-        if 'state' in data:
-            file.state = data['state']
-        
-        FileRepository.update()
+        if current_user.role == 'patient' and (datetime.utcnow() - file.creation_date > timedelta(minutes=60)):
+            return {'message': 'Update not allowed after 15 minutes'}, 403
 
-        return {'message': 'File updated successfully'}, 200
+        if current_user.role == 'patient' and file.patient_id == current_user.id:
+            file.presentation_time = new_presentation_time
+            FileRepository.update(file)
+            return {'message': 'Presentation time updated successfully'}, 200
+
+        return {'message': 'Permission denied'}, 403
 
     @staticmethod
-    def delete_file(file_id):
+    def delete_file(file_id, current_user):
         file = FileRepository.get_by_id(file_id)
         if not file:
             return {'message': 'File not found'}, 404
-        
-        FileRepository.delete(file)
 
-        return {'message': 'File deleted successfully'}, 200
+        if current_user.role == 'patient' and file.patient_id == current_user.id:
+            FileRepository.delete(file)
+            return {'message': 'File deleted successfully'}, 200
+
+        return {'message': 'Permission denied'}, 403
+    
+
+#mostrar gestion de agenda medic 
+    @staticmethod
+    def get_appointments_by_date_range(medic_id, start_date_str, end_date_str):
+        start_date = datetime.fromisoformat(start_date_str)
+        end_date = datetime.fromisoformat(end_date_str)
+        files = FileRepository.get_by_date_range(medic_id, start_date, end_date)
+        
+        return [
+            {
+                'id': fl.id,
+                'patient_name': f"{fl.patient.name} {fl.patient.last_name}",
+                'medic_name': f"{fl.medic.name} {fl.medic.last_name}",
+                'speciality': fl.medic.speciality,
+                'presentation_time': fl.presentation_time,
+                'creation_date': fl.creation_date,
+                'details': fl.details,
+                'state': fl.state
+            } for fl in files
+        ], 200
